KEEP-169 View in Explorer in AccountInfo (#177)
