+++ src/copied/_locales/en/extension.en.json
@@ -7,6 +7,7 @@
     "backUp": "Backup phrase",
     "cancel": "Cancel",
     "copied": "Copied!",
+    "copyToClipboard": "Click to copy to clipboard",
     "enter": "Enter",
     "modal": {
       "error": "Incorrect password"
@@ -16,7 +17,7 @@
     "privKey": "Private key",
     "pubKey": "Public key",
     "seed": "Backup phrase",
-    "copyToClipboard": "Click to copy to clipboard"
+    "viewInExplorer": "View in Explorer"
   },
   "assetInfo": {
     "id": "ID",
