+++ src/ui/components/pages/assets/helpers.ts
@@ -1,33 +1,2 @@
-const explorerUrls = new Map([
-  ['W', 'wavesexplorer.com'],
-  ['T', 'testnet.wavesexplorer.com'],
-  ['S', 'stagenet.wavesexplorer.com'],
-  ['custom', 'wavesexplorer.com/custom'],
-]);
-
-export function getTxHistoryLink(networkCode: string, address: string): string {
-  const explorer = explorerUrls.get(
-    explorerUrls.has(networkCode) ? networkCode : 'custom'
-  );
-  return `https://${explorer}/address/${address}/tx/`;
-}
-
-export function getTxDetailLink(networkCode: string, txId: string): string {
-  const explorer = explorerUrls.get(
-    explorerUrls.has(networkCode) ? networkCode : 'custom'
-  );
-  return `https://${explorer}/tx/${txId}`;
-}
-
-export function getAssetDetailLink(
-  networkCode: string,
-  assetId: string
-): string {
-  const explorer = explorerUrls.get(
-    explorerUrls.has(networkCode) ? networkCode : 'custom'
-  );
-  return `https://${explorer}/assets/${assetId}`;
-}
-
 export const icontains = (source, target) =>
   (source ?? '').toLowerCase().includes((target ?? '').toLowerCase());
