+++ src/ui/urls/index.ts
@@ -1,0 +1,34 @@
+const explorerUrls = new Map([
+  ['W', 'wavesexplorer.com'],
+  ['T', 'testnet.wavesexplorer.com'],
+  ['S', 'stagenet.wavesexplorer.com'],
+  ['custom', 'wavesexplorer.com/custom'],
+]);
+
+export function getAccountLink(networkCode: string, address: string) {
+  const explorer = explorerUrls.get(
+    explorerUrls.has(networkCode) ? networkCode : 'custom'
+  );
+  return `https://${explorer}/address/${address}`;
+}
+
+export function getTxHistoryLink(networkCode: string, address: string): string {
+  return `${getAccountLink(networkCode, address)}/tx/`;
+}
+
+export function getTxDetailLink(networkCode: string, txId: string): string {
+  const explorer = explorerUrls.get(
+    explorerUrls.has(networkCode) ? networkCode : 'custom'
+  );
+  return `https://${explorer}/tx/${txId}`;
+}
+
+export function getAssetDetailLink(
+  networkCode: string,
+  assetId: string
+): string {
+  const explorer = explorerUrls.get(
+    explorerUrls.has(networkCode) ? networkCode : 'custom'
+  );
+  return `https://${explorer}/assets/${assetId}`;
+}
